feat: inject import.meta.env w/ recma
