+++ packages/integrations/mdx/src/index.ts
@@ -1,26 +1,38 @@
+import type { AstroIntegration } from 'astro';
+import type { Plugin as VitePlugin } from 'vite';
 import { compile as mdxCompile } from '@mdx-js/mdx';
 import mdxPlugin, { Options as MdxRollupPluginOptions } from '@mdx-js/rollup';
-import type { AstroIntegration } from 'astro';
 import { parse as parseESM } from 'es-module-lexer';
 import { blue, bold } from 'kleur/colors';
 import { VFile } from 'vfile';
-import type { Plugin as VitePlugin } from 'vite';
-import { rehypeApplyFrontmatterExport } from './astro-data-utils.js';
-import type { MdxOptions } from './utils.js';
 import fs from 'node:fs/promises';
+import { getFileInfo, handleExtendsNotSupported, parseFrontmatter } from './utils.js';
 import {
-	getFileInfo,
+	recmaInjectImportMetaEnvPlugin,
+	rehypeApplyFrontmatterExport,
 	getRehypePlugins,
 	getRemarkPlugins,
-	handleExtendsNotSupported,
-	parseFrontmatter,
-} from './utils.js';
+} from './plugins.js';
+import { PluggableList } from '@mdx-js/mdx/lib/core.js';
 
 const RAW_CONTENT_ERROR =
 	'MDX does not support rawContent()! If you need to read the Markdown contents to calculate values (ex. reading time), we suggest injecting frontmatter via remark plugins. Learn more on our docs: https://docs.astro.build/en/guides/integrations-guide/mdx/#inject-frontmatter-via-remark-or-rehype-plugins';
 
 const COMPILED_CONTENT_ERROR =
 	'MDX does not support compiledContent()! If you need to read the HTML contents to calculate values (ex. reading time), we suggest injecting frontmatter via rehype plugins. Learn more on our docs: https://docs.astro.build/en/guides/integrations-guide/mdx/#inject-frontmatter-via-remark-or-rehype-plugins';
+
+export type MdxOptions = {
+	remarkPlugins?: PluggableList;
+	rehypePlugins?: PluggableList;
+	/**
+	 * Choose which remark and rehype plugins to inherit, if any.
+	 *
+	 * - "markdown" (default) - inherit your project’s markdown plugin config ([see Markdown docs](https://docs.astro.build/en/guides/markdown-content/#configuring-markdown))
+	 * - "astroDefaults" - inherit Astro’s default plugins only ([see defaults](https://docs.astro.build/en/reference/configuration-reference/#markdownextenddefaultplugins))
+	 * - false - do not inherit any plugins
+	 */
+	extendPlugins?: 'markdown' | 'astroDefaults' | false;
+};
 
 export default function mdx(mdxOptions: MdxOptions = {}): AstroIntegration {
 	return {
@@ -52,6 +64,7 @@ export default function mdx(mdxOptions: MdxOptions = {}): AstroIntegration {
 				const mdxPluginOpts: MdxRollupPluginOptions = {
 					remarkPlugins: await getRemarkPlugins(mdxOptions, config),
 					rehypePlugins: getRehypePlugins(mdxOptions, config),
+					recmaPlugins: [recmaInjectImportMetaEnvPlugin],
 					jsx: true,
 					jsxImportSource: 'astro',
 					// Note: disable `.md` support
